// src/app/api/models/card.ts
/* tslint:disable */
import { Colors } from './colors';
import { RelatedCards } from './related_cards';
import { CardFace } from './card_face';
import { Legality } from './legality';
import { ImageUri } from './image_uri';
export interface Card {
  id?: string;
  oracle_id?: string;
  multiverse_ids?: Array<number>;
  mtgo_id?: number;
  arena_id?: number;
  mtgo_foil_id?: number;
  uri?: string;
  scryfall_uri?: string;
  prints_search_uri?: string;
  rulings_uri?: string;
  name?: string;
  layout?: 'normal' | 'split' | 'flip' | 'transform' | 'meld' | 'leveler' | 'saga' | 'planar' | 'scheme' | 'vanguard' | 'token' | 'double_faced_token' | 'emblem' | 'augment' | 'host';
  cmc?: number;
  type_line?: string;
  oracle_text?: string;
  mana_cost?: string;
  power?: string;
  toughness?: string;
  loyalty?: string;
  life_modifier?: string;
  hand_modifier?: string;
  colors?: Array<Colors>;
  color_indicator?: Array<Colors>;
  color_identity?: Array<Colors>;
  all_parts?: Array<RelatedCards>;
  card_faces?: Array<CardFace>;
  legalities?: Legality;
  reserved?: boolean;
  edhrec_rank?: number;
  set?: string;
  set_name?: string;
  collector_number?: string;
  set_search_uri?: string;
  scryfall_set_uri?: string;
  image_uris?: ImageUri;
  highres_image?: boolean;
  reprint?: boolean;
  digital?: boolean;
  rarity?: 'common' | 'uncommon' | 'rare' | 'mythic';
  flavor_text?: string;
  artist?: string;
  illustration_id?: string;
  frame?: string;
  full_art?: boolean;
  watermark?: string;
  border_color?: 'black' | 'borderless' | 'gold' | 'silver' | 'white';
  story_spotlight_number?: number;
  story_spotlight_uri?: string;
  timeshifted?: boolean;
  colorshifted?: boolean;
  futureshifted?: boolean;
  purchase_uris?: {[key: string]: string};
  related_uris?: {[key: string]: string};
}